+++ badge/app-fight.h
@@ -155,7 +155,7 @@ static void sendRST(user *inbound);
 
 /* Game state */
 static uint8_t calc_level(uint16_t xp);
-static uint8_t xp_for_level(uint8_t level);
+static uint16_t xp_for_level(uint8_t level);
 static void changeState(fight_state);
 static void clearstatus(void);
 static uint8_t prevEnemy(void);
